// src/app/models/Feed.ts
export class Feed {
    feedId: number;
    titulo: string;
    descricaoPortal: string;
    textoDescricao: string;
    dataCadastro: Date;
    dataValidadeDe: Date;
    dataValidadeAte: Date;
    usuarioId: number;
    publico: boolean;
    link: string;
    imagem: string;
    ativo: boolean;
    eventoId: number;
    }